rename CreateVotePage to CreateVoteFOrm; add CreateVotePage component
